requirements.txt
crewai==0.186.1
crewai-tools
python-dotenv==1.0.0